requirements.txt
flask
selenium
webdriver-manager
requests
fpdf
Pillow
